pyproject.toml
[tool.poetry]
name = "pythonprojectdrf-coursework"
version = "0.1.0"
description = ""
authors = ["Nastya Foxe <Kukusiki2204@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.4.0"
requests = "^2.32.3"
redis = "^5.0.7"
celery = "^5.4.0"
python-dotenv = "^1.0.1"
django-celery-beat = "^2.6.0"
coverage = "^7.6.0"
status = "^0.2.5"
flake8 = "^7.1.0"
black = "^24.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
